Fixed settings
The addon had one set of defaults declared in the Lua file, responsible for the addon's actual behaviour. But there was a different set of defaults stored in the localisation file that governed the display of what options were enabled, causing a mismatch between user configuration and actual addon behaviour. This discrepancy would only resolve itself once the user manually toggled the affected options. This is now fixed, combined with a much more sane way of storing, loading and handling both default and saved settings.
The more time I spend on this addon the weirder it gets lol.